# Reduce object creation during substitution